Move File/Error dialogs out of content
