+++ src/components/FileListDialog/index.tsx
@@ -10,7 +10,7 @@ import {
 } from "@mui/material";
 import { isEqual } from "lodash";
 import { ReactComponent as CloseIconSvg } from "../../assets/icons/close_icon.svg";
-import GenericTable, { Column } from "../../components/GenericTable";
+import GenericTable, { Column } from "../GenericTable";
 import { FormatDate, paginateAndSort } from "../../utils";
 
 const StyledDialog = styled(Dialog)({
@@ -170,8 +170,8 @@ type Props = {
 const FileListDialog = ({ batch, onClose, open, ...rest }: Props) => {
   const [batchFiles, setBatchFiles] = useState<BatchFileInfo[]>([]);
   const [prevBatchFilesFetch, setPrevBatchFilesFetch] = useState<FetchListing<BatchFileInfo>>(null);
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [error, setError] = useState<string>(null);
+
+  const [, setError] = useState<string>(null);
   const [loading, setLoading] = useState<boolean>(false);
 
   const handleCloseDialog = () => {
